v0.1.5: Updated README to recommend local installation due to issue with CWD when using global installation
Currently there is a known issue with automated directory creation when running X-Tokenize globally. Certain directories such as the nft-metadata directory and logs directory are created where the user's terminal is currently open. This is unintended and we need to update the winston and nft-metadata deployment to place the files in the root path of the project no matter where it is installed.