Bug Fixes - getNFTokenIdFromTx, casing issue on two imports.
The introduction of the "fixNFTokenRemint" amendment introduced a breaking change to the way that NFTokenID's are calculated from a transaction. This fix includes the correct calculation to appropriately derive an NFTokenID from a mint tx.

This update also includes a fix to the casing issue on two imports. For details refer to https://github.com/x-Tokenize/X-Tokenize/issues/1#issue-1947450765